refactor backend into controllers and services
