+++ app/server/controllers/saveCustomizations.ts
@@ -1,12 +1,12 @@
 import type {
   SerializedCustomizeForm
 } from "~/types";
-import { updateOptionAndValueNames } from "../services/updateOptionAndValueNames";
-import { updateVariantsPriceStatusMedia } from "../updateVariantsPriceStatusMedia";
+import { updateOptionAndValueNames } from "./updateOptionAndValueNames";
+import { updateVariantsPriceStatusMedia } from "./updateVariantsPriceStatusMedia";
 import { setProductMetadata } from "../services/setProductMetadata";
 import { FLOWER_OPTION_NAME, FOXTAIL_NAMESPACE, PALETTE_OPTION_NAME, PRODUCT_METADATA_PRICES, SIZE_OPTION_NAME } from "~/constants";
 import { convertJsonToTypescript } from "~/jsonToTypescript";
-import { TwoWayFallbackMap } from "../models/TwoWayFallbackMap";
+import { TwoWayFallbackMap } from "../utils/TwoWayFallbackMap";
 
 export async function saveCustomizations(admin, data: SerializedCustomizeForm) {
