+++ app/server/services/updateProductOptionsAndVariants.ts
@@ -1,40 +1,25 @@
 import { UPDATE_PRODUCT_OPTION_AND_VARIANTS_QUERY } from "../graphql";
 
-export async function updateOptionAndValueNames(
+export async function updateProductOptionsAndVariants(
   admin,
-  product,
-  currentOptionName: string,
-  newDisplayName: string,
-  optionValueToNameUpdates: { [key: string]: string }
+  productId: string,
+  optionName: string,
+  optionId: string,
+  valuesToAdd,
+  idsToRemove,
+  valuesToUpdate
 ) {
-
-  const option = product.options.find(
-    (o) => o.name === currentOptionName,
-  );
-
-  // get shopifyIds
-  const updatedOptions: { [key: string]: string } = option
-    ? option.optionValues.reduce(function (variants: [{[key: string]: string }] , optionValue) {
-      variants.push({
-        id: optionValue.id,
-        name: optionValueToNameUpdates[optionValue.name] != null ? optionValueToNameUpdates[optionValue.name] : optionValue.name
-      });
-      return variants;
-    }, [])
-    : [];
-
-
 
   const updateProductOptionNameResponse = await admin.graphql(
     UPDATE_PRODUCT_OPTION_AND_VARIANTS_QUERY,
     {
       variables: {
-        productId: product.id,
-        optionName: newDisplayName,
-        optionId: option.id,
-        newValues: [],
-        oldValues: [],
-        updatedValues: updatedOptions
+        productId: productId,
+        optionName: optionName,
+        optionId: optionId,
+        newValues: valuesToAdd,
+        oldValues: idsToRemove,
+        updatedValues: valuesToUpdate
       },
     },
   );
